add todo and completed section

+++ src/components/TaskList.vue
@@ -1,7 +1,7 @@
 <template>
-  <div class="page-content page-container mb-5" id="page-content">
-    <div class="row container d-flex justify-content-center">
-      <div class="col-md-10">
+  <div class="mb-5">
+    <div class="row d-flex justify-content-center">
+      <div class="col-md-12">
         <div class="card px-3">
           <div class="card-body">
             <label for="newTask">Add New Task</label>
@@ -18,31 +18,55 @@
               </div>
             </div>
             <small class="text-danger" v-if="error">{{ error }}</small>
-            <hr>
-            <h5 class="card-title">Today's Task List</h5>
-            <div v-if="tasks">
-              <ul v-if="tasks.length" class="list-group">
-                <li v-for="task in tasks" :key="task.id"
-                    class="list-group-item list-group-item-action d-flex justify-content-between"
-                    :class="task.done?'task-done':''">
-                  <div @click="updateTask(task)">
-                    <div class="form-check">
-                      <input class="form-check-input" type="checkbox" id="inlineFormCheck" :checked="task.done">
-                    </div>
-                    <span class="ml-4">{{ task.title }}</span>
-                  </div>
-                  <div class="btn-clear" v-if="!task.done">
-                    <i class="remove fas fa-times" @click.prevent="deleteTask(task)"></i>
-                  </div>
-                </li>
-              </ul>
-              <div v-else>No Task Found</div>
-            </div>
-            <div v-else><Loading/></div>
           </div>
         </div>
       </div>
     </div>
+    <div class="row justify-content-around mt-3" v-if="tasks">
+      <div class="col col-6">
+        <div class="card card-body" v-if="tasks">
+          <h5 class="card-title">Todo Task(<span v-text="toDoTasks.length"></span>)</h5>
+          <ul v-if="toDoTasks.length" class="list-group">
+            <li v-for="task in toDoTasks" :key="task.id"
+                class="list-group-item list-group-item-action d-flex justify-content-between"
+                :class="task.done?'task-done':''">
+              <div @click="updateTask(task)">
+                <div class="form-check">
+                  <input class="form-check-input" type="checkbox" id="inlineFormCheck" :checked="task.done">
+                </div>
+                <span class="ml-4">{{ task.title }}</span>
+              </div>
+              <div class="btn-clear" v-if="!task.done">
+                <i class="remove fas fa-times" @click.prevent="deleteTask(task)"></i>
+              </div>
+            </li>
+          </ul>
+          <div v-else>No Task Available</div>
+        </div>
+      </div>
+      <div class="col col-6">
+        <div class="card card-body" v-if="tasks">
+          <h5 class="card-title">Completed Task (<span v-text="completedTasks.length"></span>)</h5>
+          <ul v-if="completedTasks.length" class="list-group">
+            <li v-for="task in completedTasks" :key="task.id"
+                class="list-group-item list-group-item-action d-flex justify-content-between task-done"
+            >
+              <div @click="updateTask(task)">
+                <div class="form-check">
+                  <input class="form-check-input" type="checkbox" :checked="task.done">
+                </div>
+                <span class="ml-4">{{ task.title }}</span>
+              </div>
+              <div class="btn-clear" v-if="!task.done">
+                <i class="remove fas fa-times" @click.prevent="deleteTask(task)"></i>
+              </div>
+            </li>
+          </ul>
+          <div v-else>No task completed yet</div>
+        </div>
+      </div>
+    </div>
+    <div v-else><Loading/></div>
   </div>
 </template>
 
@@ -61,7 +85,17 @@ export default {
     }
   },
   computed: {
-    ...mapState(['tasks'])
+    ...mapState(['tasks']),
+    completedTasks: function (){
+      return this.$store.state.tasks.filter(function(task) {
+        return task.done
+      })
+    },
+    toDoTasks: function (){
+      return this.$store.state.tasks.filter(function(task) {
+        return !task.done
+      })
+    }
   },
   methods: {
     createTask() {
@@ -84,9 +118,11 @@ export default {
       })
     },
     deleteTask(task) {
-      this.$store.dispatch('deleteTask', {
-        task: task
-      })
+      if(confirm('Are you sure?')) {
+        this.$store.dispatch('deleteTask', {
+          task: task
+        })
+      }
     }
   }
 }
@@ -109,8 +145,8 @@ export default {
 
 .btn-clear {
   background: red;
-  padding: 0 10px;
+  padding: 0 6px;
   color: #fff;
-  border-radius: 3px;
+  border-radius: 12px;
 }
 </style>
